+++ NEWS.md
@@ -7,8 +7,10 @@ This property is used in HTML or LaTeX deps resolution to know
 when to be recursive 
 - `dim_pretty()` returns correct numbers when not 'inches'
 - `as_flextable.table()` now propagates `...` as expected
-- when table was on two pages, there were duplicated caption entries,
-this is fixed thanks to Christophe Dervieux and Nick Bart.
+- pdf: when table was on two pages, there were duplicated caption entries,
+this is fixed thanks to Christophe Dervieux and Nick Bart. Repeating 
+the caption along pages can be desactivated with command 
+`set_table_properties(opts_pdf = list(caption_repeat = FALSE))`.
 
 ## Changes